Feature/ra/add Authorication into Review & ReviewReply

+++ src/main/java/ultimatum/project/service/review/ReviewReplyService.java
@@ -2,14 +2,16 @@ package ultimatum.project.service.review;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ultimatum.project.domain.dto.reviewReplyDTO.*;
+import ultimatum.project.domain.entity.member.Member;
 import ultimatum.project.domain.entity.review.Review;
 import ultimatum.project.domain.entity.review.ReviewReply;
-import ultimatum.project.dto.reviewReplyDTO.*;
 import ultimatum.project.global.exception.CustomException;
 import ultimatum.project.global.exception.ErrorCode;
+import ultimatum.project.repository.MemberRepository;
 import ultimatum.project.repository.ReviewReplyRepository;
 import ultimatum.project.repository.ReviewRepository;
 
@@ -24,19 +26,29 @@ public class ReviewReplyService {
 
     private final ReviewReplyRepository reviewReplyRepository;
     private final ReviewRepository reviewRepository;
+    private final MemberRepository memberRepository;
 
     @Transactional
-    public CreateReplyResponse createReply(Long reviewId, CreateReplyRequest request) {
+    public CreateReplyResponse createReply(Authentication authentication, Long reviewId, CreateReplyRequest request) {
 
+        if(authentication == null || !authentication.isAuthenticated()){
+            throw new CustomException(ErrorCode.BAD_REQUSET_USER);
+        }
+        String email = authentication.getName();
+        Member member = memberRepository.findByMemberEmail(email);
 
         //리뷰 엔티티 조회
        Review review = reviewRepository.findById(reviewId)
                 .orElseThrow(() -> new CustomException(ErrorCode.RESOURCE_NOT_FOUND));
 
+       if( member == null) {
+           throw new CustomException(ErrorCode.INVALID_MEMBER);
+       }
+
         //DTO를 엔티티로 변환 후 저장
         ReviewReply reviewReply = new ReviewReply();
         reviewReply.setReview(review);
-        reviewReply.setReviewReplyer(request.getReviewReplyer());
+        reviewReply.setReviewReplyer(email);
         reviewReply.setReviewReplyContent(request.getReviewReplyContent());
         reviewReply = reviewReplyRepository.save(reviewReply);
 
@@ -44,7 +56,7 @@ public class ReviewReplyService {
         //저장된 엔티티를 다시 DTO로 변환하여 반환
         return new CreateReplyResponse(
                 reviewReply.getReviewReplyId(),
-                reviewReply.getReviewReplyer(),
+                email,
                 reviewReply.getReviewReplyContent());
 
     }
@@ -67,11 +79,21 @@ public class ReviewReplyService {
 
 
     @Transactional
-    public UpdateReplyResponse updateReply(Long reply_id, UpdateReplyRequest request){
+    public UpdateReplyResponse updateReply(Authentication authentication, Long reply_id, UpdateReplyRequest request){
+
+        if (authentication == null || !authentication.isAuthenticated())  {
+            throw new CustomException(ErrorCode.BAD_REQUSET_USER);
+        }
+        String email = authentication.getName();
+        Member member = memberRepository.findByMemberEmail(email);
 
         ReviewReply reviewReply = reviewReplyRepository.findById(reply_id)
                         .orElseThrow(() -> new CustomException(ErrorCode.RESOURCE_NOT_FOUND));
-        reviewReply.update(request.getReviewReplyer(), request.getReviewReplyContent());
+
+        if (!member.getMemberEmail().equals(reviewReply.getReviewReplyer())) {
+            throw new CustomException(ErrorCode.REPLY_NOT_YOURS);
+        }
+        reviewReply.update(request.getReviewReplyContent());
 
         reviewReplyRepository.save(reviewReply);
 
@@ -84,10 +106,22 @@ public class ReviewReplyService {
 
 
     @Transactional
-    public DeleteReplyResponse deleteReply (Long reply_id){
+    public DeleteReplyResponse deleteReply (Authentication authentication, Long reply_id){
+
+        if(authentication == null || !authentication.isAuthenticated()){
+            throw new CustomException(ErrorCode.BAD_REQUSET_USER);
+        }
+
+        String email = authentication.getName();
+        Member member = memberRepository.findByMemberEmail(email);
 
         ReviewReply reviewReply = reviewReplyRepository.findById(reply_id)
                 .orElseThrow(()-> new CustomException(ErrorCode.RESOURCE_NOT_FOUND));
+
+        if(!member.getMemberEmail().equals(reviewReply.getReviewReplyer())){
+            throw new CustomException(ErrorCode.REPLY_NOT_YOURS);
+        }
+
         reviewReplyRepository.delete(reviewReply);
         return new DeleteReplyResponse(reply_id);
     }
